refactor: move TestGetSelfConsensusState and TestValidateSelfClient to new file
